Greatly refactored and streamlined the codebase.
The two biggest changes were implementing gin for the parameter file reading
and using the subclassing API for defining the neural network and updating
the weights. This removed a lot of internal plumbing. Also used tensorflow
to do all the image processing, which eliminated the need for skimage as a
dependency and greatly simplified the image processing part of the code.