Fixed a bug in qbrain.learn().
In the Bellman equation in qbrain.py:learn(), self.learningRate was
being used erroneously. self.learningRate applies to the optimizer's
learning rate. The variable that should be used in the Bellman equation
is self.discountRate. This was fixed. Additionally, the learning rate,
optimizer name, and loss function name were added as attributes to
the base brain class.